Extracted getDrivingDistance to separate file
